Replace SDL2 with SDL3

+++ Nu/Nu/Effects/EffectDescriptor.fs
@@ -194,7 +194,7 @@ and Content =
     | Nil // first to make default value when missing
     | StaticSprite of Image : Resource * Aspects : Aspect array * Content : Content
     | AnimatedSprite of Image : Resource * CelSize : Vector2i * CelCount : int * CelRun : int * CelDelay : GameTime * Playback : Playback * Aspects : Aspect array * Content : Content
-    | TextSprite of Font : Resource * Text : string * FontSizing : int option * FontStyling : FontStyle Set * Aspects : Aspect array * Content : Content
+    | TextSprite of Font : Resource * Text : string * FontSizing : single option * FontStyling : FontStyle Set * Aspects : Aspect array * Content : Content
     | Billboard of Albedo : Resource * Roughness : Resource * Metallic : Resource * AmbientOcclusion : Resource * Emission : Resource * Normal : Resource * HeightMap : Resource * TwoSided : bool * Clipped : bool * Aspects : Aspect array * Content : Content
     | StaticModel of Resource : Resource * Clipped : bool * Aspects : Aspect array * Content : Content
     | Light3d of LightType : LightType * Aspects : Aspect array * Content : Content
